backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
openai-whisper
ffmpeg-python
python-multipart
python-multipart
tempfile